Something

+++ Bomberman/Bomberman/Bomb.cpp
@@ -1,5 +1,5 @@
 #include "Bomb.h"
 
-Bomb::Bomb(sf::Vector2f mPosition):position(mPosition) {
+Bomb::Bomb(sf::Vector2i mPosition):position(mPosition) {
 
 }
